backend/backendAPI/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
python-dotenv
pydantic[email]
motor==3.6.1
passlib[bcrypt]==1.7.4
